the first versions of jimmy's brains
turnandgo is a simple as fuck program that makes jimmy "strafe" to the
right
Jimmybrain is the "main" function
ObstacleAnalysis has all the functions that we'll use to get sensor info
from jimmy, and a ton of other useful stuff! Try no to change
obstacleAnalysis without explaining the change beforehand, kay?